requirements.txt
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.32.3
pytz==2024.1
flask==3.0.3
schedule==1.2.2
ffmpeg-python==0.2.0
python-dotenv==1.0.1
